pyproject.toml
[tool.poetry]
name = "obsidian-rag"
version = "0.1.0"
description = ""
authors = ["Kousuke Ebihara <kousuke@co3k.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.2.5"
streamlit = "^1.35.0"
langfuse = "^2.36.1"
openai = "^1.35.1"
chromadb = "^0.5.3"
langchain-community = "^0.2.5"
python-dotenv = "^1.0.1"
langchain-openai = "^0.1.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
